pyproject.toml
[tool.poetry]
name = "dmytro-bot"
version = "0.1.0"
description = ""
authors = ["<jus1stored@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
